Mix sound using floats internally.

+++ src/client/sound/sound.h
@@ -25,8 +25,8 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #endif
 
 typedef struct {
-    int     left;
-    int     right;
+    float   left;
+    float   right;
 } samplepair_t;
 
 typedef struct sfxcache_s {
@@ -68,8 +68,8 @@ typedef struct playsound_s {
 
 typedef struct channel_s {
     sfx_t       *sfx;           // sfx number
-    int         leftvol;        // 0-255 volume
-    int         rightvol;       // 0-255 volume
+    float       leftvol;        // 0.0-1.0 volume
+    float       rightvol;       // 0.0-1.0 volume
     int         end;            // end time in global paintsamples
     int         pos;            // sample position in sfx
     int         entnum;         // to allow overriding a specific sound
